+++ docs/changelog.rst
@@ -37,6 +37,7 @@ Changelog
 * :bug:`-` Aave v2 accounting for deposit interest profit and borrow payback loss should now work correctly again.
 * :bug:`6717` Fixes an issue causing the graph for some multichain assets to not be displayed correctly.
 * :bug:`6169` Handle some kraken event categories better and show them in the UI. Categories like failed withdrawals, moving funds from/to staking and credits of forked assets or delisted asset trades.
+* :bug:`-` Users will be able to import custom assets from the assets backup file.
 
 * :release:`1.30.2 <2023-09-21>`
 * :feature:`-` Improved support for importing Binance CSV files.